refactor: move functions and modules to new util module
Move the get_config_path and send_sig functions to a new util module.
Additionally move notify and expand to util as a submodule.
