+++ src/util/expand.rs
